requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
backports.tarfile==1.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docutils==0.21.2
e2b==1.1.0
e2b-code-interpreter==1.0.5
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
id==1.5.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
kiwisolver==1.4.7
langchain==0.3.19
langchain-community==0.3.17
langchain-core==0.3.36
langchain-experimental==0.3.4
langchain-text-splitters==0.3.6
langsmith==0.3.8
llama-index-agent-openai==0.4.6
llama-index-core==0.12.19
llama-index-llms-openai==0.3.20
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplot==0.1.9
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.6.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.27.1
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.20
nltk==3.9.1
numpy==1.26.4
openai==1.63.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyloco==0.0.139
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
readme_renderer==44.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.22.3
scipy==1.13.1
SecretStorage==3.3.3
SimpleWebSocketServer==0.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
statsmodels==0.14.4
streamlit==1.42.1
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
twine==6.1.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
ushlex==0.99.1
watchdog==6.0.0
websocket-client==1.8.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
